+++ lib/generators/lite/uxid/templates/install.rb
@@ -1,13 +1,21 @@
 # frozen_string_literal: true
 
 Lite::Uxid.configure do |config|
-  config.hashid_charset = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-  config.hashid_salt = 1_369_136
-  config.hashid_size = 16
-  config.nanoid_charset = "_-0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-  config.nanoid_size = 21
-  config.obfuscateid_spin = 0
-  config.ulid_charset = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-  config.ulid_size = 26
-  config.uuid_version = 4
+  # HashID
+  # config.hashid_charset = Lite::Uxid::ALPHANUMERIC
+  # config.hashid_salt = 1_369_136
+
+  # NanoID
+  # config.nanoid_charset = Lite::Uxid::WEB_SAFE
+  # config.nanoid_size = 21
+
+  # ObfuscatedID
+  # config.obfuscateid_spin = 0
+
+  # ULID
+  # config.ulid_charset = Lite::Uxid::COCKFORDS_32
+  # config.ulid_size = 26
+
+  # UUID
+  # config.uuid_version = 4
 end