+++ spec/lite/uxid/reversible/hashid_spec.rb
@@ -5,8 +5,8 @@ require "spec_helper"
 RSpec.describe Lite::Uxid::Reversible::Hashid do
   let(:prefix) { nil }
   let(:id) { 10 }
-  let(:hashid) { "1zWr1m0" }
-  let(:prefixed_hashid) { "test_1zWr1m0" }
+  let(:hashid) { "7pau2oXSklq0" }
+  let(:prefixed_hashid) { "test_7pau2oXSklq0" }
   let(:encoder) { described_class.new(id, prefix: prefix) }
   let(:decoder) { described_class.new(hashid, prefix: prefix) }
 
@@ -27,20 +27,20 @@ RSpec.describe Lite::Uxid::Reversible::Hashid do
   end
 
   describe ".encode" do
-    it 'to be "1zWr1m0"' do
+    it 'to be "7pau2oXSklq0"' do
       expect(described_class.encode(id)).to eq(hashid)
       expect(encoder.encode).to eq(hashid)
     end
 
-    it "to be 7" do
-      expect(described_class.encode(id).size).to eq(7)
-      expect(encoder.encode.size).to eq(7)
+    it "to be 12" do
+      expect(described_class.encode(id).length).to eq(12)
+      expect(encoder.encode.length).to eq(12)
     end
 
     context "with prefix" do
       let(:prefix) { "test_" }
 
-      it 'to be "test_1zWr1m0"' do
+      it 'to be "test_7pau2oXSklq0"' do
         expect(described_class.encode(id, prefix: prefix)).to eq(prefixed_hashid)
         expect(encoder.encode).to eq(prefixed_hashid)
       end