+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2] - 2024-09-23
+### Changed
+- Improve configuration template
+- Improve hashid generation
+### Removed
+- Hashid size option
+
 ## [2.0.1] - 2024-09-23
 ### Changed
 - Renamed `Scatterid` to `Obfuscateid`